feat(currency-debt): LOT 7 — flux financiers numeric + 4 séquelles d'arrondi corrigées
Les 4 colonnes de flux financier les plus exposées du reliquat après le
LOT 6 (43 fichiers chacune) : refunds.amount_kmf, store_credits.
amount_kmf/remaining_kmf, cash_collections.amount_kmf,
cash_deposits.amount_kmf.

Ce sont des montants réellement remis ou dus à une personne : un
remboursement, un avoir, de l'argent liquide compté. Les arrondir à
l'unité fait perdre des centimes à quelqu'un, à chaque opération.

store_credits.remaining_kmf est converti EN MÊME TEMPS qu'amount_kmf
délibérément : un avoir dont le montant porte des centimes mais dont le
solde ne peut pas les représenter produirait un résidu non consommable
à chaque usage partiel — même défaut total/lignes corrigé au LOT 6.

Aucune vue, aucun trigger, aucune contrainte ne dépend de ces colonnes
(vérifié par pg_depend/pg_trigger/pg_constraint sur base réellement
migrée). Lot le plus simple du chantier sur ce plan.

QUATRE SÉQUELLES D'ARRONDI corrigées, trouvées en cherchant les
parseInt/Math.round appliqués à des colonnes DÉJÀ converties par les
lots précédents — des conversions qui annulaient silencieusement le
bénéfice de la migration :

- services/invoice-service.js : Math.round(price_kmf) sur les lignes
  d'une FACTURE CLIENT. order_items.price_kmf est numeric depuis la
  migration 219 (LOT 6) ; cet arrondi falsifiait les montants facturés.
  Le plus grave des quatre.
- routes/finance.js : parseInt(total_kmf) dans l'export de
  rapprochement comptable Stripe — parseFloat était déjà utilisé pour
  total_eur juste au-dessus, l'incohérence saute aux yeux une fois vue.
- routes/parcel-label.js, routes/logistics.js : parseInt(total_kmf) sur
  les totaux imprimés (étiquette colis, manifeste PDF).

Ces quatre sites illustrent une limite des lots précédents : convertir
la colonne ne suffit pas si le code la re-tronque à la lecture. Aucun
gate ne détecte ce motif — piste pour un futur garde-fou.

Validé par mutation réelle : refunds.amount_kmf et
store_credits.remaining_kmf remis en integer -> 3 tests échouent ;
migration réappliquée -> 5/5 vert, zéro résidu.

Contrainte métier découverte à l'exécution : idx_cash_coll_order impose
UN encaissement par commande — le test d'agrégation crée donc une
commande par montant plutôt que d'en réutiliser une seule.

Reliquat après ce lot : 15 colonnes integer sur 120 (était 20).
Restent principalement customs_history/shipments (douane, 0-1 fichier),
fabrics, unsold_items, disputes.refund_kmf, cart_shares.cart_total_kmf.
Cas à part maintenu hors périmètre : exchange_rates.aed_kmf/eur_kmf
sont des TAUX, pas des montants — problème distinct et conceptuellement
plus grave, à traiter séparément.

Gates verts : check-currency-format, feature-guard, code-quality-gate,
backend:audit. Suite unitaire : 75 échecs, identique à la baseline
mesurée avant ce lot malgré 4 fichiers source modifiés — aucune
régression.

+++ routes/finance.js
@@ -194,7 +194,11 @@ router.get('/stripe-proofs', ...adminOnly, async (req, res, next) => {
         komerce_reference:  order.reference,
         stripe_payment_id:  order.stripe_payment_id,
         amount_eur:         parseFloat(order.total_eur),
-        amount_kmf:         parseInt(order.total_kmf),
+        // parseFloat, pas parseInt : orders.total_kmf est numeric(14,2)
+        // depuis la migration 213 (chantier currency debt). parseInt
+        // tronquait les centimes dans cet export de rapprochement comptable
+        // Stripe — séquelle de l'époque où la colonne était integer.
+        amount_kmf:         parseFloat(order.total_kmf),
         created_at:         order.created_at,
         client_name:        order.client_name,
         client_email:       order.client_email,
